web/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.1
python-dotenv==1.0.0
email-validator==2.0.0.post2 